reorganize to allow for multiple log queries
