+++ user_count.py
@@ -1,6 +1,7 @@
+#!/usr/bin/env python
 from fetch import fetch
 
 log_types = ["spu", "auditfile"]
 fetch_count = fetch.Fetch(log_types)
 
-from count import users
+from users import count
